// src/(FSD)/shareds/types/pharmacys/PharmacyInfo.type.ts
export interface PharmacyInfoType {
    phId: number;
    phName: string;
    phTel: string;
    phAdd: string;
    timeWeekStartDate: string;
    timeWeekEndDate: string;
    timeSatStartDate: string;
    timeSatEndDate: string;
    timeHoliStartDate: string;
    timeHoliEndDate: string;
    enjoyIndex: number;
    starAvg: number;
    reviewIndex: number;
    enjoy: boolean;
}